+++ app/src/main/res/layout/fragment_reportranking_report.xml
@@ -16,15 +16,15 @@
         android:layout_height="match_parent"
         android:orientation="vertical">
 
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rcv_ReportShelf"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:paddingLeft="25dp"
-            android:paddingTop="15dp"
-            android:paddingRight="25dp"
-            android:paddingBottom="10dp"
-            android:scrollbars="vertical" />
+<!--        <androidx.recyclerview.widget.RecyclerView-->
+<!--            android:id="@+id/rcv_ReportShelf"-->
+<!--            android:layout_width="match_parent"-->
+<!--            android:layout_height="match_parent"-->
+<!--            android:paddingLeft="25dp"-->
+<!--            android:paddingTop="15dp"-->
+<!--            android:paddingRight="25dp"-->
+<!--            android:paddingBottom="10dp"-->
+<!--            android:scrollbars="vertical" />-->
 
     </RelativeLayout>
