--- app/src/main/res/layout/item_mail.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:gravity="center"
-    android:background="@drawable/border"
-    android:padding="10dp">
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_vertical"
-        android:gravity="center_vertical">
-
-        <de.hdodenhof.circleimageview.CircleImageView
-            android:id="@+id/iv_profileImg"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:layout_marginLeft="10dp"
-            android:layout_gravity="center_vertical"
-            android:src="@drawable/icon_logo"
-            app:civ_border_color="@color/black"
-            app:civ_border_width="1dp" />
-
-
-
-        <RelativeLayout
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_vertical"
-            android:layout_marginLeft="10dp"
-            android:orientation="vertical"
-            android:textAlignment="gravity">
-
-            <TextView
-                android:id="@+id/tv_nickname"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/font_main"
-                android:text="nickname"
-                android:textColor="@color/black"
-                android:textSize="12sp"
-                android:textStyle="bold" />
-
-
-            <TextView
-                android:id="@+id/tv_time"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="시간"
-                android:layout_marginLeft="230dp"
-                android:fontFamily="@font/font_main"
-                android:textSize="12sp" />
-
-            <TextView
-                android:id="@+id/tv_mailtext"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="쪽지내용"
-                android:layout_below="@id/tv_nickname"
-                android:layout_marginTop="10dp"
-                android:fontFamily="@font/font_main"
-                android:textSize="10sp"
-                android:textStyle="bold" />
-
-        </RelativeLayout>
-
-
-
-    </LinearLayout>
-
-
-</LinearLayout>
-
-
